refactor(exceptions): Moved LexerError into the new exceptions.py file.
